requirements.txt
# Core dependencies
psycopg2-binary==2.9.7
pandas==2.0.3
geopandas==0.13.2
sqlalchemy==2.0.19
requests==2.31.0

# Data processing
numpy==1.24.3
shapely==2.0.1
fiona==1.9.4

# Database tools
python-dotenv==1.0.0
configparser==6.0.0

# Visualization & Analysis
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Optional: Advanced features
streamlit==1.25.0
jupyter==1.0.0
scikit-learn==1.3.0
torch==2.0.1
transformers==4.32.0 